fix: mark: recall the position properly and add tests
fix: substitute: replace unscaled & with the name of the matched text

fix: substitute: if the replacement text is simply "%" then replace it with the
previous non empty replacement text.

fix: set dirty flag for all the commands that affect the buffer

I have also decided to make most of the editor structs variables
lowercase, this will make them private or unexported because there is no
reason for them to be public.