fix: when checking the change, skip some of the commands that do not need it
commands such as write, edit and quit don't rely on the start and end position
being valid so there is no point of checking that the range is valid.

fix: don't print the file size when writing or reading if the suppress
diagnostics flag is enabled.

while here, add docs to the missing functions