feat: implement a basic undo system
The way it works is for every supported command it will record a
tiny operation of that command in to an array. So deleting 5 lines
will record 5 sub tasks of inserting the line. When the command is
considered "done" and has finished executing it'll push that array
to an internal array. Executing the "u" command after to undo the
previous action will traverse the internal array and execute each
"sub task" in a reverse order. The following commands are supported
and therefore undo-able: a, c, d, i, j, m, r and t. I am missing
undo actions for the global commands (g, v, G and V). Executing an
undo command will not create a redo task.