fix off by one when executing global commands
Before executing the command on the marked lines we have to keep
track of the how many lines there are before doing anything and how
many there are in the buffer currently. For every iteration of the
loop we subtract the dot with the buffer count and the current
count to get the correct dot.