wrap the editor error in a `explainError` when executing `h`
when executing the `h` command it does not modify `printErrors`
which would result in the error defaulting to `ErrDefault`, so to
counteract that we would just print the error directly to the err
printer. The problem with that is that the main loop prints any
errors returned so it would result in two different errors, the
actual error and `ErrDefault`.

Now doing something like this will work as expected:

	$ ./ed file
	1365
	}
	?
	h
	unknown command
	h
	unknown command
	d
	h

Notice that it does not print `?` and then the actual `invalid
command` error like before. Also note that the last `h` does not
print the previous error, because the `d` command was succesful: