fix address calculation within the substitute command
The for loop within Editor.substitute() always started
with the first line (ed.lines[0]), not the line referred to
by ed.start (ed.lines[ed.start-1]).
This patch adjusts the for loop header so that `i' gets set to
the correct line indices.

The expected buffer and output for the "s/.*/some/" test case
has been adjusted, since the command now works on the last line,
as dot is set to the last line after loading the text;
before it worked on the first line, since substitute's for
loop always started with ed.lines[0].

An additional test case has been added: "3,5s/ ./X/"
It tests whether substitute works at the expected address range.

Part of #1.