fix submatch index calculation within substitute
'submatch' is the result of a call to re.FindAllStringSubmatch.
The first entry, submatch[0][0], is the match of the
entire expression. The first submatch, which \1 refers to,
is in submatch[0][1]. Therefore it is not necessary to
subtract 1 from n before accessing the slice.

The test case for "3,5s/ ./X/" has been extended to also test
for correct handling of replacement expressions \1 and \2.

Fixes #1.